+++ lib/membrane/core/parent/clock_handler.ex
@@ -10,6 +10,10 @@ defmodule Membrane.Core.Parent.ClockHandler do
           Core.Parent.state()
         ) ::
           Core.Parent.state() | no_return
+  def choose_clock(_children, nil, state) do
+    state
+  end
+
   def choose_clock(children, provider, state) do
     %{synchronization: synchronization} = state
 
@@ -20,17 +24,13 @@ defmodule Membrane.Core.Parent.ClockHandler do
       end
 
     components = components ++ children
-
-    cond do
-      provider != nil -> set_clock_provider(get_clock_from_provider(components, provider), state)
-      synchronization.clock_provider.choice == :manual -> state
-      true -> choose_clock_provider(components) |> set_clock_provider(state)
-    end
+    clock = get_clock_from_provider(components, provider)
+    set_clock_provider(clock, state)
   end
 
   @spec reset_clock(Core.Parent.state()) :: Core.Parent.state()
   def reset_clock(state),
-    do: set_clock_provider(%{clock: nil, provider: nil, choice: :auto}, state)
+    do: set_clock_provider(%{clock: nil, provider: nil}, state)
 
   defp set_clock_provider(clock_provider, state) do
     Clock.proxy_for(state.synchronization.clock_proxy, clock_provider.clock)
@@ -48,22 +48,7 @@ defmodule Membrane.Core.Parent.ClockHandler do
         raise ParentError, "#{inspect(provider)} is not a clock provider"
 
       %{clock: clock} ->
-        %{clock: clock, provider: provider, choice: :manual}
-    end
-  end
-
-  defp choose_clock_provider(components) do
-    case components |> Enum.filter(& &1.clock) do
-      [] ->
-        %{clock: nil, provider: nil, choice: :auto}
-
-      [%{name: name, clock: clock}] ->
-        %{clock: clock, provider: name, choice: :auto}
-
-      components ->
-        raise ParentError, """
-        Cannot choose clock for the parent, as multiple components provide one, namely: #{Enum.map_join(components, ", ", & &1.name)}. Please explicitly select the clock by setting `ChildrenSpec.clock_provider` parameter.
-        """
+        %{clock: clock, provider: provider}
     end
   end
 end