+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.1
  * Specify the order in which state fields will be printed in the error logs. [#614](https://github.com/membraneframework/membrane_core/pull/614)
  * Handle buffers, only if demand on input pad with `flow_control: :auto` is non-negative. [#654](https://github.com/membraneframework/membrane_core/pull/654)
+ * Fix clock selection [#626](https://github.com/membraneframework/membrane_core/pull/626)
 
 ## 1.0.0
  * Introduce `:remove_link` action in pipelines and bins.